@@ -12,7 +12,7 @@ authors = [
 ]
 description = "Vease-Viewer is the viewer microservice of Vease"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9, <3.13"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
